Move testOutputUtils.ts to correct testutils directory
